section 5 (partial): extract StagedEdit + apply to model
New src/model/ tree with edit_queue.rs (195 lines): owns the StagedEdit
enum, the resolve_dir_by_path walk helper, and a pure apply_edit dispatch.
browse_view.rs::apply_staged_edits shrinks from ~165 lines to a 25-line
view-side wrapper.

The Vec<StagedEdit> field and its many query helpers (pending_adds_for,
find_pending_add_idx, resolved_hfs_type_creator, etc.) stay on BrowseView
for now — folding those into an EditQueue model object is a separate
(larger) iteration.
