+++ src/model/mod.rs
@@ -1,0 +1,6 @@
+//! Model layer: non-GUI state and business logic.
+//!
+//! The eventual goal is for every GUI tab to be a renderer over an object in
+//! this tree. See §5 / §10 of `docs/codecleanup.md`.
+
+pub mod edit_queue;
